+++ org.eclipse.winery.frontends/app/tosca-management/src/app/instance/sharedComponents/propertiesDefinition/propertiesDefinition.component.html
@@ -54,37 +54,31 @@
 
         <div *ngIf="propertiesDefinitions.selectedValue === propertiesEnum.Custom">
             <tabset *wineryRepositoryHideOnFeature="configEnum.Yaml">
-                <tab heading="Properties">
-                    <winery-dynamic-table [dynamicMetadata]="dynamicTableData"
-                                          tableTitle="Properties"
-                                          [modalTitle]="modalTitle"
-                                          [data]="this.propertiesDefinitions.winerysPropertiesDefinition.propertyDefinitionKVList"
-                                          [avoidDuplicateProperties]="['key']"
-                                          (entryAdded)="onChangeProperty()"
-                                          (entryEdited)="onChangeProperty()"
-                                          (entryRemoved)="onChangeProperty()"></winery-dynamic-table>
+                <tab heading="Property Definitions">
+                    
+                    <br/>
 
                     <button (click)="show.inherited = !show.inherited"> {{ show.inherited ? "Hide Inheritance Details" : "Show Inheritance Details" }}</button>
-                    
+
                     <div [hidden]="!show.inherited">
-                        
+
                         <br/>
 
                         <alert type="info" *ngIf="!this.inheritedPropertiesDefinitions.length">
                             There is no inheritance.
                         </alert>
-                        
+
                         <alert type="info" *ngIf="this.inheritedPropertiesDefinitions.length > 1">
-                            The order of the tables are the same as in the inheritance order.
-                            The highest table corresponds to the highest parent in the inheritance order.
+                            The order of the tables are the same as in the inheritance hierarchy.
+                            The highest table corresponds to the highest parent in the inheritance hierarchy.
                         </alert>
-                        
+
                         <ng-container *ngFor="let parent of this.inheritedPropertiesDefinitions.reverse(); index as index">
-                            
+
                             <hr *ngIf="index !== 0"/>
 
                             <winery-dynamic-table
-                                [tableTitle]="'Properties derived from ' + parent.parent"
+                                [tableTitle]="'Property Definitions from ' + parent.parent"
                                 [titleRouterLink]="getParentRouterLink(parent.parent)"
                                 [dynamicMetadata]="dynamicTableData"
                                 [data]="parent.properties.winerysPropertiesDefinition.propertyDefinitionKVList"
@@ -92,9 +86,22 @@
                                 [disableButtons]="true"
                             >
                             </winery-dynamic-table>
-
+                            
                         </ng-container>
+
+                        <hr *ngIf="this.inheritedPropertiesDefinitions"/>
+
                     </div>
+                    
+                    <winery-dynamic-table [dynamicMetadata]="dynamicTableData"
+                                          tableTitle="Effective Property Definitions (including Inheritance)"
+                                          [modalTitle]="modalTitle"
+                                          [data]="this.propertiesDefinitions.winerysPropertiesDefinition.propertyDefinitionKVList"
+                                          [avoidDuplicateProperties]="['key']"
+                                          (entryAdded)="onChangeProperty()"
+                                          (entryEdited)="onChangeProperty()"
+                                          (entryRemoved)="onChangeProperty()"></winery-dynamic-table>
+                    
 
                 </tab>
                 <tab heading="Wrapper">
