+++ org.eclipse.winery.model/org.eclipse.winery.model.tosca.canonical/src/main/java/org/eclipse/winery/model/tosca/extensions/kvproperties/PropertyDefinitionKV.java
@@ -20,6 +20,8 @@ import java.util.Objects;
 
 import javax.xml.bind.annotation.XmlElement;
 
+import org.eclipse.winery.model.tosca.utils.ModelUtilities;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -123,26 +125,13 @@ public class PropertyDefinitionKV implements Serializable {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         PropertyDefinitionKV that = (PropertyDefinitionKV) o;
-        return Objects.equals(key, that.key);
-    }
-
-    // TODO: add additional equals to equals function above or does that break other logic?
-    public boolean equalsAllProperties(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        PropertyDefinitionKV that = (PropertyDefinitionKV) o;
-
-        // Handle the problem "null vs empty list"
-        List<ConstraintClauseKV> effectiveConstraintList = that.constraintList;
-        if (effectiveConstraintList == null) effectiveConstraintList = new ArrayList<>();
         
-        return Objects.equals(key, that.key) 
+        return Objects.equals(key, that.key)
             && Objects.equals(type, that.type)
             && Objects.equals(required, that.required)
             && Objects.equals(defaultValue, that.defaultValue)
             && Objects.equals(description, that.description)
-            && constraintList.containsAll(effectiveConstraintList)
-            && constraintList.size() == effectiveConstraintList.size();
+            && ModelUtilities.compareUnorderedNullableLists(constraintList, that.constraintList);
     }
 
     @Override
